# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CalorieDiet Swagger Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SERVER"
        }
    ],
    "paths": {
...
        "/api/food/index": {
            "get": {
                "tags": [
                    "Food"
                ],
                "summary": "Get index of food",
                "operationId": "2e0343a7c0609576e8c8fcff2da8a1e8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/food/all": {
            "get": {
                "tags": [
                    "Food"
                ],
                "summary": "Get list of food",
                "operationId": "268f6495a5e2005223a207a4a654e60c",
                "parameters": [
                    {
                        "name": "food_category_id",
                        "in": "query",
                        "description": "food_category_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/ticket-categories/index": {
            "get": {
                "tags": [
                    "Ticket categories"
                ],
                "summary": "Get list of ticket categories",
                "operationId": "2120df4a7a6d085eff90bc00a33cf72d",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}